broke meta tag functions into separate files
