+++ client/src/utils/determineContentTypeFromExtension.js
@@ -1,0 +1,21 @@
+const determineContentTypeFromExtension = (thumbnail) => {
+  if (thumbnail) {
+    const fileExt = thumbnail.substring(thumbnail.lastIndexOf('.'));
+    switch (fileExt) {
+      case 'jpeg':
+      case 'jpg':
+        return 'image/jpg';
+      case 'png':
+        return 'image/png';
+      case 'gif':
+        return 'image/gif';
+      case 'mp4':
+        return 'video/mp4';
+      default:
+        return '';
+    }
+  }
+  return '';
+};
+
+module.exports = determineContentTypeFromExtension;
